Use individual read and write transfers
'Interface Description Sensirion SCD30 Sensor Module' states in section
1.4.9 'Read firmware version' that this requires individual write and
read transfer and this returns meaningful data for the first time.

It is not clear to me, why the I2C protocol description in this data
sheet mentions combined transfers with a repeated start condition when
they are not used/supported by the sensor commands.